requirements.txt
asgiref==3.11.1
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.2
djangorestframework==3.16.1
flake8==7.3.0
idna==3.11
mccabe==0.7.0
mysqlclient==2.2.8
numpy==2.4.2
pandas==3.0.0
pycodestyle==2.14.0
pyflakes==3.4.0
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
tabulate==0.9.0
urllib3==2.6.3
